+++ src/commands/osu_scores.ts
@@ -323,7 +323,7 @@ export const maplb = async (input: bottypes.commandInput) => {
             }
             const temp = helper.tools.commands.getButtonArgs(input.id);
             if (temp.error) {
-                input.interaction.reply({
+                input.interaction.followUp({
                     content: helper.vars.errors.paramFileMissing,
                     ephemeral: true,
                     allowedMentions: { repliedUser: false }
@@ -1179,7 +1179,7 @@ export const recent = async (input: bottypes.commandInput) => {
             commanduser = input.interaction.member.user;
             const temp = helper.tools.commands.getButtonArgs(input.id);
             if (temp.error) {
-                input.interaction.reply({
+                input.interaction.followUp({
                     content: helper.vars.errors.paramFileMissing,
                     ephemeral: true,
                     allowedMentions: { repliedUser: false }
@@ -2604,7 +2604,7 @@ export const scores = async (input: bottypes.commandInput) => {
             commanduser = input.interaction.member.user;
             const temp = helper.tools.commands.getButtonArgs(input.id);
             if (temp.error) {
-                input.interaction.reply({
+                input.interaction.followUp({
                     content: helper.vars.errors.paramFileMissing,
                     ephemeral: true,
                     allowedMentions: { repliedUser: false }